postgres_to_es/requirements.txt
python-dotenv==1.0.0
psycopg2-binary==2.9.9
requests==2.31.0
pydantic-settings==2.1.0